fix: 123file does not represent a valid word 'file'
A valid 'word' must start with 4 continuous characters, folloed by
letters or numbers. String '123file' is not a word. In the last
version, I mistakenly thought that '123file' could be understood as
a valid word 'file'. In fact, '123file' is a state which does not
start with 4 English letters, must get a 'Separator' before judging
it as a valid word. The separator is a space or a non-numeric, non-
alphabetic character. For example 'fi123,file' can be recognize a
valid word 'file' because there is a comma separator.